requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv==1.0.0
openai>=1.86.0
jinja2==3.1.2
